First working version of points2Obj_mesh
Adjacent fibres are stitched together into a series of quads

Shown to work on a really simple mesh.

Needs general review.

The files size of the obj is big. Point previously pushed into
the vertex array are not reused. ( When a vertex could be reused 4 times)

Needs testing to make sure it meshes tricky funnels not just the nice curved forms.

Needs clippy fixes.