Added bevy_plugin -- initial version
Sphere on the left will be a gizmo of defining a
line-segment on a sphere.

Mesh on the right is the corresponding hopf-mesh.

This first version demonstrates the rendering of a sample mesh.

It looks like there is a bug in mesh generation ... It look like
the first band is repeated.