Various small refinements.
I was worried about the correctness of the generated normals
So I created a new Mesh(object) made of lines which show the normals

Now following the convention of calling .build() to return a Mesh
The API MeshBuilder::build() is limited in that it can only panic

So I inserted a intermediate step 'HopfMeshBuilder::construct()' which
follows a better Error reporting strategy.

Next, Dropped a hard coded const N_TRIES it can now be configured in the builder.

Also, Now supplying a per vertex uv's.

Various clippy documentation fixes.