Fibre::build() - Improved numerical stability in the adaptive loop
Previously when I tried alpha in the less than  [0..4PI]
I saw the target_sample was not always met or exceeded.
This is fundamental to a quad based mesh - adjacent fibres are
stitched together.

For example alpha [0..3PI] with target_samples set to 80
I would see 81 points returned.

The solution is to tighten the tolerance as the fibre length decreases

This leads to the seeming contradiction that, it may takes more compute,
more iterations of the adaptive loop, for fibre segment with alpha[0..2PI],
 than for a longer fibre segment.

Added test fibre_build_tight_tolerance() for the range [0..2PI]

NB [0..PI] is still unstable.