Reduced the memory footprint, when writing OBJ files.
Previous Vertex's were written to both a vertex_store and a vertex_buffer

The store is a HashMap used for duplicating.

I deleted the vertex buffer, so all the duplicate array filling is removed.

But this comes at the computational expense of recreating the vertex buffer from the HashMap
in the final write() call.

NB write now comsumes the OBJ struct.