1) New: SurfacePoint{lat, lon} because I want to be clear out what is a latitude and what is a longgitude.
2) *BREAKING CHANGE* Convert all workspace items from f64 to f32.

I was following the convention that all of the fibre computation
was a f64.

In bevy all objects are f32 forcing a convesion at this point.
I have simplified by systematically using f32 throughout.

3) debug_asserts now have error messages.

4) Worked on bevy_hopf - Added Handles to the sphere that defines
the hopf mesh.

I think the arrangement of observer handles ok
Making the handles children on the ball seems a good move.

I am checking in so I have a reference for debugging.